changed Login.vue to Auth.vue
